+++ src/data/usecases/download-subtitle/external-download-subtitle-protocols.ts
@@ -1,0 +1,3 @@
+export { DownloadSubtitle } from '../../../domain/usecases/download-subtitle'
+export { CaptionModel } from '../../../domain/models/caption'
+export { DownloadSubtitleService } from '../../protocols/download-subtitle-service'
